DBZ-8956 move tests in deployment
