+++ quarkus-debezium-parent/quarkus-debezium-postgres-parent/deployment/pom.xml
@@ -59,9 +59,6 @@
                 <artifactId>maven-surefire-plugin</artifactId>
                 <version>3.5.2</version>
                 <configuration>
-                    <includes>
-                        <include>**/EngineProcessorTest.java</include>
-                    </includes>
                     <systemPropertyVariables>
                         <java.util.logging.manager>org.jboss.logmanager.LogManager</java.util.logging.manager>
                         <maven.home>${maven.home}</maven.home>
